Add newest localization keys to reflect changes in rewards page
A new version of the rewards page was released on September 24, 2019
including a long list of improvements for both the user interface as
well as the back-end server. Among these changes were the addition of
new localization keys to translate new elements in the user interface
to the currently supported languages.